refactor xml records path
